requirements.txt
# Core libraries for video/audio processing
moviepy==1.0.3
gTTS
pyttsx3
edge-tts
matplotlib
Pillow==10.1.0
elevenlabs>=1.0.0 
whisper-timestamped


# API clients and utilities
google-api-python-client
google-auth-oauthlib
groq
requests
python-dotenv

# LangChain/LangGraph for future agentic orchestration
langchain
langgraph

# Vector database for knowledge base
chromadb

# YouTube metadata scraping
yt-dlp

# HuggingFace datasets for story training data
datasets

# Reddit API
praw
